put constant in a separate file
