+++ src/app/core/constants.ts
@@ -1,0 +1,1 @@
+export const HUNDRED_TB = 109951162777600;
